+++ codigo/fps.c++
@@ -1,10 +1,10 @@
 #include "fps.h++"
 #include <iostream>
 
-unsigned int Fps::Nanosegundos = 0;
-unsigned int Fps::Nanosegundos_fotogramas = 0;
+long int Fps::Nanosegundos = 0;
+long int Fps::Nanosegundos_fotogramas = 0;
 bool Fps::Mostrar_fps = true;
-unsigned int Fps::Contador_fps = 0;
+long int Fps::Contador_fps = 0;
 
 Tiempo Fps::Tiempo_actual = Reloj::now();
 Tiempo Fps::Tiempo_anterior = Reloj::now();
@@ -19,13 +19,13 @@ unsigned int Fps::Calcular_tiempo()
 
 	if(Fps::Nanosegundos_fotogramas >= 250000000)//Mostrar cada 250 milisegundos
 	{
-		unsigned int fps_total = Fps::Contador_fps + 1;
+		long int fps_total = Fps::Contador_fps + 1;
 
 		Fps::Contador_fps = 0;
 		Fps::Tiempo_anterior_fotogramas = Fps::Tiempo_actual;
 		Fps::Mostrar_fps = true;
 
-		return 1000000000.0/(static_cast<double>(Fps::Nanosegundos_fotogramas)/static_cast<double>(fps_total));//Retorna FPS
+		return static_cast<unsigned int>(1000000000.0/(static_cast<double>(Fps::Nanosegundos_fotogramas)/static_cast<double>(fps_total)));//Retorna FPS
 	}
 	else
 	{
@@ -42,5 +42,5 @@ bool Fps::Actualizar_fps()
 
 unsigned int Fps::Obtener_nanosegundos()
 {
-	return Fps::Nanosegundos;
+	return static_cast<unsigned int>(Fps::Nanosegundos);
 }
